[feat] create separate file for back button components.
