+++ src/Components/BackButtons/styles.ts
@@ -1,0 +1,22 @@
+import { StyleSheet } from 'react-native';
+
+import { colors } from '../../styles/colors';
+
+export default StyleSheet.create({
+  backButtonContainer: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+    columnGap: 10,
+  },
+  headerText: {
+    color: colors.midRed,
+    textAlign: 'center',
+    fontSize: 12,
+    fontStyle: 'normal',
+    fontWeight: '400',
+    lineHeight: 24,
+    letterSpacing: -0.264,
+  },
+});
